requirements.txt
langchain
ibm-generative-ai
python-dotenv
